Move BackupFilePair to it's own file, and implement missing properties
