+++ ValheimBackupShared/BO/Backup.cs
@@ -130,60 +130,4 @@ namespace ValheimBackup.BO
             return "-" + dt.ToString("yyyyMMddHHmmss");
         }
     }
-
-    [JsonObject(MemberSerialization.OptIn)]
-    public class BackupFilePair : INotifyPropertyChanged
-    {
-        private string _sourcePath;
-        private string _destinationPath;
-        private string _name;
-        private string _timeHash;
-        private string _extension;
-
-        [JsonProperty]
-        public string SourcePath
-        {
-            get => _sourcePath;
-            set
-            {
-                if (_sourcePath != value)
-                {
-                    _sourcePath = value;
-                    NotifyPropertyChanged("SourcePath");
-                }
-            }
-        }
-
-        [JsonProperty]
-        public string DestinationPath
-        {
-            get => _destinationPath;
-            set
-            {
-                if (_destinationPath != value)
-                {
-                    _destinationPath = value;
-                    NotifyPropertyChanged("DestinationPath");
-                }
-            }
-        }
-
-        public BackupFilePair() { }
-
-        public BackupFilePair(string source, string destination, string name, string extension, string timeHash)
-        {
-            _sourcePath = source;
-            _destinationPath = destination;
-        }
-
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        private void NotifyPropertyChanged(string propertyName)
-        {
-            if (PropertyChanged != null)
-            {
-                PropertyChanged(this, new PropertyChangedEventArgs(propertyName));
-            }
-        }
-    }
 }
